+++ vp9/encoder/vp9_tokenize.c
@@ -257,7 +257,8 @@ static void tokenize_b(int plane, int block, BLOCK_SIZE plane_bsize,
 
     add_token(&t, coef_probs[band[c]][pt],
               vp9_dct_value_tokens_ptr[v].extra,
-              vp9_dct_value_tokens_ptr[v].token, skip_eob,
+              (uint8_t)vp9_dct_value_tokens_ptr[v].token,
+              (uint8_t)skip_eob,
               counts[band[c]][pt]);
     eob_branch[band[c]][pt] += !skip_eob;
